+++ fsl_mrs/scripts/fsl_dynmrs.py
@@ -51,9 +51,21 @@ def main():
                               help='limit the fit optimisation to a chemical shift range. '
                                    'Defaults to a nucleus-specific range. '
                                    'For 1H default=(.2,4.2).')
-    fitting_args.add_argument('--baseline_order', default=2, type=int,
+    fitting_args.add_argument('--baseline',
+                              type=str,
+                              default='poly, 2',
+                              help='Select baseline type. '
+                                   'Options: "off", "polynomial", or "spline". '
+                                   'With "polynomial" specify an order, e.g. "polynomial, 2". '
+                                   'With spline specify a stiffness: '
+                                   "'very-stiff', 'stiff', 'moderate', 'flexible', and 'very-flexible', "
+                                   "or an effective dimension (2 -> inf) where 2 is the stiffest. "
+                                   "The default is 'polynomial, 2'.")
+    fitting_args.add_argument('--baseline_order',
+                              default=None,
+                              type=int,
                               metavar=('ORDER'),
-                              help='order of baseline polynomial (default=2).')
+                              help=configargparse.SUPPRESS)
     fitting_args.add_argument('--metab_groups', default=0, nargs='+',
                               type=str_or_int_arg,
                               help='metabolite groups: list of groups'
@@ -250,6 +262,7 @@ def main():
 
     # Fitting Arguments
     Fitargs = {'ppmlim': args.ppmlim,
+               'baseline': args.baseline,
                'baseline_order': args.baseline_order,
                'metab_groups': metab_groups,
                'model': 'voigt'